refactor: extract useTakeScreenshot hook to separate file to avoid dependency cycles and clean up MobileBrowserBottomBar component
