+++ packages/kit/src/views/Discovery/components/MobileBrowser/MobileBrowserBottomBar.tsx
@@ -4,7 +4,4 @@ function MobileBrowserBottomBar(_props: IMobileBrowserBottomBarProps) {
   return null;
 }
 
-export const useTakeScreenshot = (_id?: string | null) => () =>
-  Promise.resolve(true);
-
 export default MobileBrowserBottomBar;
