+++ packages/kit/src/views/Discovery/hooks/useTakeScreenshot.ts
@@ -1,0 +1,3 @@
+// This stub is used on web/desktop/extensions where the screenshot feature is not needed.
+export const useTakeScreenshot = (_id?: string | null) => () =>
+  Promise.resolve(true);
